[UNSTABLE SUCK] trying to separate schema
